# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Proyectos y Canales",
    "version": "1.0.0",
    "description": "API para gestionar proyectos, canales y autenticación"
  },
  "paths": {
...
    "/api/category": {
      "get": {
        "summary": "Obtiene las categorías de un proyecto",
        "tags": ["Category"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Texto de búsqueda para filtrar categorías por nombre"
          },
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del proyecto al que pertenecen las categorías"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Número de página para paginación"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Cantidad de categorías por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías recuperada exitosamente"
          },
          "400": {
            "description": "Se requiere projectId"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Proyecto no encontrado o sin acceso"
          },
          "500": {
            "description": "Error al recuperar categorías"
          }
        }
      },
      "post": {
        "summary": "Crea una nueva categoría en un proyecto",
        "tags": ["Category"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectId": {
                    "type": "string",
                    "description": "ID del proyecto donde se creará la categoría"
                  },
                  "name": {
                    "type": "string",
                    "description": "Nombre de la categoría"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada exitosamente"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Proyecto no encontrado o sin permisos suficientes"
          },
          "500": {
            "description": "Error al crear la categoría"
          }
        }
      },
      "put": {
        "summary": "Actualiza una categoría existente",
        "tags": ["Category"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID de la categoría a actualizar"
                  },
                  "name": {
                    "type": "string",
                    "description": "Nuevo nombre de la categoría"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada exitosamente"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tienes permiso para editar esta categoría"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "500": {
            "description": "Error al actualizar la categoría"
          }
        }
      },
      "delete": {
        "summary": "Elimina una categoría",
        "tags": ["Category"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "description": "ID de la categoría a eliminar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría eliminada exitosamente"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tienes permiso para eliminar esta categoría"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "500": {
            "description": "Error al eliminar la categoría"
          }
        }
      }
    },
...
  }
}